Update login form with success state; add success and error colors to styles
The login form has been updated to include a success state. Success and error colors have been added to the styles.

The `index.html` file has been modified to change the class of the input group to "input-group success" when there are no errors.

The `styles.css` file has been updated to add new variables for success and error colors, and to apply the success color as a box-shadow to the input field when there are no errors. The primary dark color has also been updated to have a higher saturation.